backend/requirements.txt
# Core Flask dependencies
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
Flask-SocketIO==5.4.1
Flask-JWT-Extended==4.6.0
Flask-Session==0.5.0

# Database and migrations
SQLAlchemy==2.0.36
Alembic==1.14.0
greenlet==3.2.3

# Security
Werkzeug==3.1.1
python-dotenv==1.0.1
bcrypt==4.2.1
Flask-WTF==1.2.1
WTForms==3.2.1

# File processing
openpyxl==3.1.5
pandas==2.2.3
Pillow==11.0.0

# PDF Generation
WeasyPrint==63.1
cffi==1.17.1

# Redis and caching
redis==5.2.1
hiredis==3.1.0

# WebSocket support
python-socketio==5.12.0
python-engineio==4.11.0

# HTTP requests and API
httpx==0.28.1
requests==2.32.3

# AI and OpenAI (optional)
openai==1.56.2
tiktoken==0.8.0

# Date and time
python-dateutil==2.9.0.post0
pytz==2024.2

# Validation and serialization
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
jsonschema==4.23.0

# Development and testing
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
coverage==7.6.9
black==24.10.0
flake8==7.1.1
isort==5.13.2
pre-commit==4.0.1

# Production server
gunicorn==23.0.0
waitress==3.0.2

# Logging and monitoring
structlog==24.4.0
sentry-sdk[flask]==2.19.2

# Environment and configuration
python-decouple==3.8
click==8.1.8

# API documentation
flasgger==0.9.7.1
Flask-RESTX==1.3.0

# Utilities
faker==33.1.0
celery==5.4.0
